requirements.txt
streamlit
langchain-core
langchain-google-genai
langgraph
pandas
openpyxl
xlrd
python-dotenv
